adding authorisation on '/api/contacts' route

+++ server/src/controller/contactController.ts
@@ -4,15 +4,15 @@ import Contact from '../models/contact.model';
 
 // @desc GET all contacts
 // @route GET /api/contacts
-// @access public
+// @access private
 export const getContacts = asyncHandler(async (req: Request, res: Response) => {
-    const contacts = await Contact.find({});
+    const contacts = await Contact.find({ user_id: req.user.id });
     res.status(200).json({ message: contacts });
 });
 
 // @desc Create new contacts
 // @route POST /api/contacts
-// @access public
+// @access private
 export const createNewContact = async (req: Request, res: Response) => {
     const { name, email, phone } = req.body;
     if (!name) {
@@ -28,7 +28,12 @@ export const createNewContact = async (req: Request, res: Response) => {
         throw new Error(`phone field is missing`);
     }
 
-    const newContact = new Contact({ name, email, phone });
+    const newContact = new Contact({
+        name,
+        email,
+        phone,
+        user_id: req.user.id,
+    });
     await newContact.save();
     res.status(201).json({ message: newContact });
     try {
@@ -42,13 +47,21 @@ export const createNewContact = async (req: Request, res: Response) => {
 
 // @desc Update all contacts
 // @route PUT /api/contacts
-// @access public
+// @access private
 export const updateContact = asyncHandler(
     async (req: Request, res: Response) => {
         const reqContact = await Contact.findById(req.params.id);
         if (!reqContact) {
             res.status(404);
         }
+
+        if (String(reqContact?.user_id) !== req.user.id) {
+            res.status(403);
+            throw new Error(
+                "User doesn't have permission to update other user"
+            );
+        }
+
         const updatedContact = await Contact.findByIdAndUpdate(
             req.params.id,
             req.body,
@@ -65,6 +78,14 @@ export const deleteContact = asyncHandler(
             res.status(404);
             throw new Error('NO contact found');
         }
+
+        if (String(reqContact?.user_id) !== req.user.id) {
+            res.status(403);
+            throw new Error(
+                "User doesn't have permission to update other user"
+            );
+        }
+
         await Contact.deleteOne({ _id: req.params.id });
         res.status(200).json({ message: `/delete contact: ${req.params.id}` });
     }
